requirements.txt
attrs==24.2.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2